+++ tests/integration/bedrock_opensearch.py
