+++ tests/integration/haystack_opensearch_sample.py
