+++ tests/integration/haystack_sample.py
