+++ tests/integration/langchain_sample_blob.py
