+++ tests/unit/app_hosting_types_tests.py
