+++ tests/unit/embeddings_wrapper.py
