+++ tests/unit/http_span_exporter.py
