+++ tests/unit/langchain_kwargs_test.py
@@ -6,17 +6,12 @@ import time
 
 import unittest
 from unittest.mock import ANY, MagicMock, patch
-from urllib.parse import urlparse
-import pytest
 import requests
-from dummy_class import DummyClass
-from langchain_openai import AzureOpenAI
-from embeddings_wrapper import HuggingFaceEmbeddings
-from http_span_exporter import HttpSpanExporter
+from monocle.tests.unit.embeddings_wrapper import HuggingFaceEmbeddings
+from monocle.tests.unit.http_span_exporter import HttpSpanExporter
 from langchain.prompts import PromptTemplate
 from langchain.schema import StrOutputParser
 from langchain_community.vectorstores import faiss
-from langchain_core.messages.ai import AIMessage
 from langchain_core.runnables import RunnablePassthrough
 from monocle_apptrace.instrumentation.common.span_handler import WORKFLOW_TYPE_MAP
 from monocle_apptrace.instrumentation.common.constants import (
@@ -34,20 +29,18 @@ from monocle_apptrace.instrumentation.common.instrumentor import (
     set_context_properties,
     setup_monocle_telemetry,
 )
-from monocle_apptrace.instrumentation.common.wrapper_method import WrapperMethod
 from opentelemetry import trace
 from opentelemetry.sdk.resources import SERVICE_NAME, Resource
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
 
-from fake_list_llm import FakeListLLM
+from monocle.tests.unit.fake_list_llm import FakeListLLM
 from parameterized import parameterized
 from monocle_apptrace.instrumentation.common.span_handler import SpanHandler
-from monocle_apptrace.instrumentation.common.wrapper import task_wrapper
 
 logger = logging.getLogger()
 logger.setLevel(logging.DEBUG)
-fileHandler = logging.FileHandler('traces.txt' ,'w')
+fileHandler = logging.FileHandler('../traces.txt', 'w')
 formatter = logging.Formatter('%(asctime)s [%(levelname)s] %(name)s: %(message)s')
 fileHandler.setFormatter(formatter)
 logger.addHandler(fileHandler)
@@ -71,7 +64,7 @@ class TestHandler(unittest.TestCase):
     def __format_docs(self, docs):
         return "\n\n ".join(doc.page_content for doc in docs)
 
-    def __createChain(self, llm_type):
+    def __createChain(self):
 
         resource = Resource(attributes={
             SERVICE_NAME: "coffee_rag_fake"
@@ -85,24 +78,14 @@ class TestHandler(unittest.TestCase):
         self.instrumentor = MonocleInstrumentor(handlers=SpanHandler())
         self.instrumentor.instrument()
         self.processor = monocleProcessor
-        responses =[self.ragText]
-        if llm_type == "FakeListLLM":
-            responses = [self.ragText]
-            llm = FakeListLLM(responses=responses)
-            llm.api_base = "https://example.com/"
-        else:
-            llm = AzureOpenAI(
-                azure_deployment=os.environ.get("AZURE_OPENAI_API_DEPLOYMENT"),
-                api_key=os.environ.get("AZURE_OPENAI_API_KEY"),
-                api_version=os.environ.get("AZURE_OPENAI_API_VERSION"),
-                azure_endpoint=os.environ.get("AZURE_OPENAI_ENDPOINT"),
-                temperature=0.1,
-                model="gpt-3.5-turbo-0125"
-            )
-            llm.azure_endpoint = "https://example.com/"
+
+        responses = [self.ragText]
+        llm = FakeListLLM(responses=responses)
+        llm.api_base = "https://example.com/"
+
         embeddings = HuggingFaceEmbeddings(model_id = "multi-qa-mpnet-base-dot-v1")
         my_path = os.path.abspath(os.path.dirname(__file__))
-        model_path = os.path.join(my_path, "./vector_data/coffee_embeddings")
+        model_path = os.path.join(my_path, "../vector_data/coffee_embeddings")
         vectorstore = faiss.FAISS.load_local(model_path, embeddings, allow_dangerous_deserialization = True)
 
         retriever = vectorstore.as_retriever()
@@ -116,10 +99,7 @@ class TestHandler(unittest.TestCase):
         return rag_chain
 
     def setUp(self):
-        os.environ["AZURE_OPENAI_API_DEPLOYMENT"] = "AZURE_OPENAI_API_DEPLOYMENT"
-        os.environ["AZURE_OPENAI_API_KEY"] = "AZURE_OPENAI_API_KEY"
-        os.environ["AZURE_OPENAI_ENDPOINT"] = "AZURE_OPENAI_ENDPOINT"
-        os.environ["AZURE_OPENAI_API_VERSION"] = "2024-02-01"
+
         os.environ["HTTP_API_KEY"] = "key1"
         os.environ["HTTP_INGESTION_ENDPOINT"] = "https://localhost:3000/api/v1/traces"
 
@@ -128,9 +108,12 @@ class TestHandler(unittest.TestCase):
         return super().tearDown()
 
     @parameterized.expand([
-        ("AzureOpenAI", AZURE_ML_ENDPOINT_ENV_NAME, "AzureOpenAI"),
-        ("FakeListLLM", AZURE_ML_ENDPOINT_ENV_NAME, "FakeListLLM"),
+        ("1", AZURE_ML_ENDPOINT_ENV_NAME, AZURE_ML_SERVICE_NAME),
+        ("2", AZURE_FUNCTION_WORKER_ENV_NAME, AZURE_FUNCTION_NAME),
+        ("3", AZURE_APP_SERVICE_ENV_NAME, AZURE_APP_SERVICE_NAME),
+        ("4", AWS_LAMBDA_ENV_NAME, AWS_LAMBDA_SERVICE_NAME),
     ])
+
     @patch.object(requests.Session, 'post')
     def test_llm_chain(self, test_name, test_input_infra, llm_type, mock_post):
         app_name = "test"
@@ -149,7 +132,7 @@ class TestHandler(unittest.TestCase):
             context_value = "context_value_1"
             set_context_properties({context_key: context_value})
 
-            self.chain = self.__createChain(llm_type)
+            self.chain = self.__createChain()
             mock_post.return_value.status_code = 201
             mock_post.return_value.json.return_value = 'mock response'
 
@@ -170,23 +153,17 @@ class TestHandler(unittest.TestCase):
             root_attributes = [x for x in dataJson["batch"] if x["parent_id"] == "None"][0]["attributes"]
             assert root_attributes["entity.1.name"] == app_name
             assert root_attributes["entity.1.type"] == WORKFLOW_TYPE_MAP['langchain']
-            if llm_type == "FakeListLLM":
-                llm_vector_store_retriever_span = [x for x in dataJson["batch"] if 'langchain_core.vectorstores.base.VectorStoreRetriever' in x["name"]][0]
-                inference_span = [x for x in dataJson["batch"] if 'fake_list_llm.FakeListLLM' in x["name"]][0]
 
-                assert llm_vector_store_retriever_span["attributes"]["span.type"] == "retrieval"
-                assert llm_vector_store_retriever_span["attributes"]["entity.1.name"] == "FAISS"
-                assert llm_vector_store_retriever_span["attributes"]["entity.1.type"] == "vectorstore.FAISS"
-                assert inference_span['attributes']["entity.1.inference_endpoint"] == "https://example.com/"
-            else:
-                llm_azure_openai_span = [x for x in dataJson["batch"] if 'langchain_openai.llms.azure.AzureOpenAI' in x["name"]][0]
-                assert llm_azure_openai_span["attributes"]["span.type"] == "inference"
-                assert llm_azure_openai_span["attributes"]["entity.1.type"] == "inference.azure_oai"
-                # assert llm_azure_openai_span["attributes"]["entity.1.provider_name"] == urlparse(os.environ.get("AZURE_OPENAI_ENDPOINT")).hostname
-                assert llm_azure_openai_span["attributes"]["entity.1.deployment"] == os.environ.get("AZURE_OPENAI_API_DEPLOYMENT")
-                assert llm_azure_openai_span["attributes"]["entity.1.inference_endpoint"] == "https://example.com/"
-                assert llm_azure_openai_span["attributes"]["entity.2.type"] == "model.llm.gpt-3.5-turbo-0125"
-                assert "Latte is a coffee drink" in llm_azure_openai_span["events"][1]['attributes']['response'][0]
+            llm_vector_store_retriever_span = [x for x in dataJson["batch"] if 'langchain_core.vectorstores.base.VectorStoreRetriever' in x["name"]][0]
+            inference_span = [x for x in dataJson["batch"] if 'FakeListLLM' in x["name"]][0]
+
+            assert llm_vector_store_retriever_span['attributes']['entity.1.name'] == "FAISS"
+            assert llm_vector_store_retriever_span['attributes']['entity.1.type'] == "vectorstore.FAISS"
+
+            # using kwargs for provider name and inference endpoint in metamodel
+            # assert inference_span['attributes']['entity.1.provider_name'] == "example.com"
+            assert inference_span['attributes']["entity.1.inference_endpoint"] == "https://example.com/"
+
 
         finally:
             os.environ.pop(test_input_infra)
