+++ tests/unit/langchain_workflow_name_test.py
@@ -1,5 +1,3 @@
-from unittest import IsolatedAsyncioTestCase
-import unittest
 
 import json
 import logging
@@ -8,42 +6,49 @@ import time
 
 import unittest
 from unittest.mock import ANY, MagicMock, patch
-from unittest import IsolatedAsyncioTestCase
 
 import requests
-from dummy_class import DummyClass
-from embeddings_wrapper import HuggingFaceEmbeddings
-from http_span_exporter import HttpSpanExporter
+
+from monocle.tests.unit.embeddings_wrapper import HuggingFaceEmbeddings
+from monocle.tests.unit.http_span_exporter import HttpSpanExporter
 from langchain.prompts import PromptTemplate
 from langchain.schema import StrOutputParser
 from langchain_community.vectorstores import faiss
-from langchain_core.messages.ai import AIMessage
 from langchain_core.runnables import RunnablePassthrough
+from monocle_apptrace.instrumentation.common.constants import (
+    AZURE_APP_SERVICE_ENV_NAME,
+    AZURE_APP_SERVICE_NAME,
+    AZURE_FUNCTION_NAME,
+    AZURE_FUNCTION_WORKER_ENV_NAME,
+    AZURE_ML_ENDPOINT_ENV_NAME,
+    AZURE_ML_SERVICE_NAME,
+    AWS_LAMBDA_ENV_NAME,
+    AWS_LAMBDA_SERVICE_NAME
+)
 from monocle_apptrace.instrumentation.common.instrumentor import (
     MonocleInstrumentor,
     set_context_properties,
     setup_monocle_telemetry,
 )
 
-from monocle_apptrace.instrumentation.common.wrapper_method import WrapperMethod
 from opentelemetry import trace
 from opentelemetry.sdk.resources import SERVICE_NAME, Resource
 from opentelemetry.sdk.trace import TracerProvider
 from opentelemetry.sdk.trace.export import BatchSpanProcessor, ConsoleSpanExporter
-from monocle_apptrace.instrumentation.common.wrapper import atask_wrapper, task_wrapper
-from fake_list_llm import FakeListLLM
+
+from monocle.tests.unit.fake_list_llm import FakeListLLM
+from parameterized import parameterized
 from monocle_apptrace.instrumentation.common.span_handler import SpanHandler
 
 logger = logging.getLogger()
 logger.setLevel(logging.DEBUG)
-fileHandler = logging.FileHandler('traces.txt', 'w')
+fileHandler = logging.FileHandler('../traces.txt', 'w')
 formatter = logging.Formatter('%(asctime)s [%(levelname)s] %(name)s: %(message)s')
 fileHandler.setFormatter(formatter)
 logger.addHandler(fileHandler)
-events = []
 
+class TestWorkflowEntityProperties(unittest.TestCase):
 
-class Test(IsolatedAsyncioTestCase):
     prompt = PromptTemplate.from_template(
         """
         <s> [INST] You are an assistant for question-answering tasks. Use the following pieces of retrieved context
@@ -72,21 +77,21 @@ class Test(IsolatedAsyncioTestCase):
 
         traceProvider.add_span_processor(monocleProcessor)
         trace.set_tracer_provider(traceProvider)
-        self.instrumentor = MonocleInstrumentor(handlers=SpanHandler())
+        self.instrumentor = MonocleInstrumentor(handlers=SpanHandler)
         self.instrumentor.instrument()
         self.processor = monocleProcessor
-        responses = [self.ragText]
+        responses =[self.ragText]
         llm = FakeListLLM(responses=responses)
         llm.api_base = "https://example.com/"
-        embeddings = HuggingFaceEmbeddings(model_id="multi-qa-mpnet-base-dot-v1")
+        embeddings = HuggingFaceEmbeddings(model_id = "multi-qa-mpnet-base-dot-v1")
         my_path = os.path.abspath(os.path.dirname(__file__))
-        model_path = os.path.join(my_path, "./vector_data/coffee_embeddings")
-        vectorstore = faiss.FAISS.load_local(model_path, embeddings, allow_dangerous_deserialization=True)
+        model_path = os.path.join(my_path, "../vector_data/coffee_embeddings")
+        vectorstore = faiss.FAISS.load_local(model_path, embeddings, allow_dangerous_deserialization = True)
 
         retriever = vectorstore.as_retriever()
 
         rag_chain = (
-                {"context": retriever | self.__format_docs, "question": RunnablePassthrough()}
+                {"context": retriever| self.__format_docs, "question": RunnablePassthrough()}
                 | self.prompt
                 | llm
                 | StrOutputParser()
@@ -94,14 +99,21 @@ class Test(IsolatedAsyncioTestCase):
         return rag_chain
 
     def setUp(self):
-        events.append("setUp")
-
-    async def asyncSetUp(self):
         os.environ["HTTP_API_KEY"] = "key1"
         os.environ["HTTP_INGESTION_ENDPOINT"] = "https://localhost:3000/api/v1/traces"
 
+
+    def tearDown(self) -> None:
+        return super().tearDown()
+
+    @parameterized.expand([
+        ("1", AZURE_ML_ENDPOINT_ENV_NAME, AZURE_ML_SERVICE_NAME),
+        ("2", AZURE_FUNCTION_WORKER_ENV_NAME, AZURE_FUNCTION_NAME),
+        ("3", AZURE_APP_SERVICE_ENV_NAME, AZURE_APP_SERVICE_NAME),
+        ("4", AWS_LAMBDA_ENV_NAME, AWS_LAMBDA_SERVICE_NAME),
+    ])
     @patch.object(requests.Session, 'post')
-    async def test_response(self, mock_post):
+    def test_llm_chain(self, test_name, test_input_infra, test_output_infra, mock_post):
         app_name = "test"
         wrap_method = MagicMock(return_value=3)
         setup_monocle_telemetry(
@@ -109,16 +121,10 @@ class Test(IsolatedAsyncioTestCase):
             span_processors=[
                 BatchSpanProcessor(HttpSpanExporter("https://localhost:3000/api/v1/traces"))
             ],
-            wrapper_methods=[
-                WrapperMethod(
-                    package="langchain.chat_models.base",
-                    object_name="BaseChatModel",
-                    method="invoke",
-                    wrapper_method=atask_wrapper
-                )
-
-            ])
+            wrapper_methods=[])
         try:
+
+            os.environ[test_input_infra] = "1"
             context_key = "context_key_1"
             context_value = "context_value_1"
             set_context_properties({context_key: context_value})
@@ -132,7 +138,7 @@ class Test(IsolatedAsyncioTestCase):
             assert response == self.ragText
             time.sleep(5)
             mock_post.assert_called_with(
-                url='https://localhost:3000/api/v1/traces',
+                url = 'https://localhost:3000/api/v1/traces',
                 data=ANY,
                 timeout=ANY
             )
@@ -140,22 +146,37 @@ class Test(IsolatedAsyncioTestCase):
             '''mock_post.call_args gives the parameters used to make post call.
             This can be used to do more asserts'''
             dataBodyStr = mock_post.call_args.kwargs['data']
-            dataJson = json.loads(dataBodyStr)  # more asserts can be added on individual fields
+            dataJson =  json.loads(dataBodyStr) # more asserts can be added on individual fields
 
-            llm_span = [x for x in dataJson["batch"] if "FakeListLLM" in x["name"]][0]
+            root_span = [x for x in dataJson["batch"] if x["parent_id"] == "None"][0]
 
-            assert llm_span["attributes"]["span.type"] == "inference"
-            # assert llm_span["attributes"]["entity.1.provider_name"] == "example.com"
-            assert llm_span["attributes"]["entity.1.type"] == "inference.azure_oai"
-            assert llm_span["attributes"]["entity.1.inference_endpoint"] == "https://example.com/"
+            # workflow_name and workflow_type in new format entity.{index}.name and entity.{index}.type
+
+            assert root_span["attributes"]["entity.1.name"] == "test"
+            assert root_span["attributes"]["entity.1.type"] == "workflow.langchain"
+            # input_found = False
+            # output_found = False
+
+            # for event in root_span['events']:
+            #     if event['name'] == "data.input" and event['attributes']['input'] == query:
+            #         input_found = True
+            #     elif event['name'] == "data.output" and event['attributes']['response'] == self.ragText:
+            #         output_found = True
+            #
+            # assert input_found
+            # assert output_found
+
 
         finally:
+            os.environ.pop(test_input_infra)
             try:
-                if self.instrumentor is not None:
+                if(self.instrumentor is not None):
                     self.instrumentor.uninstrument()
             except Exception as e:
                 print("Uninstrument failed:", e)
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
     unittest.main()
+
+
