+++ tests/unit/vectorstore_deployment_test.py
